+++ TeleSharp.TL/TL/Messages/TLRequestToggleChatAdmins.cs
@@ -1,10 +1,5 @@
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TeleSharp.TL;
+
 namespace TeleSharp.TL.Messages
 {
     [TLObject(-326379039)]
@@ -18,7 +13,7 @@ namespace TeleSharp.TL.Messages
             }
         }
 
-        public int ChatId { get; set; }
+        public long ChatId { get; set; }
         public bool Enabled { get; set; }
         public TLAbsUpdates Response { get; set; }
 
@@ -30,7 +25,7 @@ namespace TeleSharp.TL.Messages
 
         public override void DeserializeBody(BinaryReader br)
         {
-            ChatId = br.ReadInt32();
+            ChatId = br.ReadInt64();
             Enabled = BoolUtil.Deserialize(br);
 
         }
