+++ TeleSharp.TL/TL/TLUpdateEncryptedMessagesRead.cs
@@ -1,10 +1,5 @@
-using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TeleSharp.TL;
+
 namespace TeleSharp.TL
 {
     [TLObject(956179895)]
@@ -18,7 +13,7 @@ namespace TeleSharp.TL
             }
         }
 
-        public int ChatId { get; set; }
+        public long ChatId { get; set; }
         public int MaxDate { get; set; }
         public int Date { get; set; }
 
@@ -30,10 +25,9 @@ namespace TeleSharp.TL
 
         public override void DeserializeBody(BinaryReader br)
         {
-            ChatId = br.ReadInt32();
+            ChatId = br.ReadInt64();
             MaxDate = br.ReadInt32();
             Date = br.ReadInt32();
-
         }
 
         public override void SerializeBody(BinaryWriter bw)
